feat(suite-native): make token selectors network agnostic
