+++ suite-native/module-accounts-management/src/screens/AccountDetailContentScreen.tsx
@@ -8,7 +8,7 @@ import {
 } from '@suite-common/wallet-core';
 import { TokenAddress } from '@suite-common/wallet-types';
 import { EventType, analytics } from '@suite-native/analytics';
-import { selectEthereumAccountTokenInfo } from '@suite-native/tokens';
+import { selectAccountTokenInfo } from '@suite-native/tokens';
 import { Screen } from '@suite-native/navigation';
 import { TransactionList } from '@suite-native/transactions';
 
@@ -34,7 +34,7 @@ export const AccountDetailContentScreen = ({
     );
 
     const token = useSelector((state: AccountsRootState) =>
-        selectEthereumAccountTokenInfo(state, accountKey, tokenContract),
+        selectAccountTokenInfo(state, accountKey, tokenContract),
     );
 
     useEffect(() => {
