+++ suite-native/receive/src/screens/ReceiveModalScreen.tsx
@@ -24,7 +24,7 @@ import {
     selectAccountNetworkSymbol,
     selectDeviceAccountKeyForNetworkSymbolAndAccountTypeWithIndex,
 } from '@suite-common/wallet-core';
-import { selectEthereumAccountTokenSymbol } from '@suite-native/tokens';
+import { selectAccountTokenSymbol } from '@suite-native/tokens';
 import { CryptoIcon } from '@suite-common/icons-deprecated';
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
 
@@ -70,8 +70,8 @@ const ReceiveModalScreenSubHeader = ({ accountKey, tokenContract }: ScreenSubHea
     const networkSymbol = useSelector((state: AccountsRootState) =>
         selectAccountNetworkSymbol(state, accountKey),
     );
-    const ethereumTokenSymbol = useSelector((state: AccountsRootState) =>
-        selectEthereumAccountTokenSymbol(state, accountKey, tokenContract),
+    const tokenSymbol = useSelector((state: AccountsRootState) =>
+        selectAccountTokenSymbol(state, accountKey, tokenContract),
     );
 
     useEffect(() => {
@@ -98,7 +98,7 @@ const ReceiveModalScreenSubHeader = ({ accountKey, tokenContract }: ScreenSubHea
                         {accountLabel && (
                             <Text variant="highlight">
                                 {accountLabel}
-                                {ethereumTokenSymbol && ` - ${ethereumTokenSymbol}`}
+                                {tokenSymbol && ` - ${tokenSymbol}`}
                             </Text>
                         )}
                     </HStack>
