+++ suite-native/state/tsconfig.json
@@ -34,6 +34,7 @@
         { "path": "../module-send" },
         { "path": "../settings" },
         { "path": "../storage" },
+        { "path": "../tokens" },
         {
             "path": "../../packages/transport-native"
         },
