+++ suite-native/tokens/package.json
@@ -16,7 +16,6 @@
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
-        "@suite-common/wallet-utils": "workspace:*",
         "@trezor/blockchain-link": "workspace:*",
         "proxy-memoize": "2.0.2",
         "react": "18.2.0"
