+++ suite-native/tokens/src/index.ts
@@ -1,4 +1,3 @@
 export * from './types';
 export * from './utils';
-export * from './ethereumTokensSelectors';
 export * from './tokensSelectors';
