+++ suite-native/tokens/src/tokensSelectors.ts
@@ -1,70 +1,249 @@
-import { TokenDefinitionsRootState } from '@suite-common/token-definitions';
-import { NetworkSymbol } from '@suite-common/wallet-config';
-import { AccountsRootState, DeviceRootState, selectAccountByKey } from '@suite-common/wallet-core';
+import { A, G, pipe } from '@mobily/ts-belt';
+import { memoizeWithArgs } from 'proxy-memoize';
 
 import {
-    selectAnyOfTokensIsKnown,
-    selectHasDeviceAnyEthereumTokens,
-    selectNumberOfEthereumAccountTokensWithFiatRates,
-} from './ethereumTokensSelectors';
+    selectFilterKnownTokens,
+    selectIsSpecificCoinDefinitionKnown,
+    TokenDefinitionsRootState,
+} from '@suite-common/token-definitions';
+import {
+    AccountsRootState,
+    DeviceRootState,
+    selectAccountByKey,
+    selectAccountTransactions,
+    selectVisibleDeviceAccountsByNetworkSymbol,
+    TransactionsRootState,
+} from '@suite-common/wallet-core';
+import {
+    AccountKey,
+    TokenAddress,
+    TokenInfoBranded,
+    TokenSymbol,
+} from '@suite-common/wallet-types';
+import { TokenInfo, TokenTransfer } from '@trezor/blockchain-link';
+import { NetworkSymbol } from '@suite-common/wallet-config';
+
 import { isCoinWithTokens } from './utils';
+import { TypedTokenTransfer, WalletAccountTransaction } from './types';
 
 export type TokensRootState = AccountsRootState & DeviceRootState & TokenDefinitionsRootState;
 
-export const selectHasDeviceAnyTokens = (state: TokensRootState, coin: NetworkSymbol) => {
-    if (!isCoinWithTokens(coin)) return false;
+export const selectAccountTokenInfo = memoizeWithArgs(
+    (
+        state: AccountsRootState,
+        accountKey?: AccountKey,
+        tokenAddress?: TokenAddress,
+    ): TokenInfoBranded | null => {
+        const account = selectAccountByKey(state, accountKey);
+        if (!account || !account.tokens) {
+            return null;
+        }
 
-    switch (coin) {
-        case 'eth':
-            const hasAnyTokens = selectHasDeviceAnyEthereumTokens(state);
+        return (
+            (A.find(
+                account.tokens,
+                (token: TokenInfo) => token.contract === tokenAddress,
+            ) as TokenInfoBranded) ?? null
+        );
+    },
+    // 100 is a reasonable size for the cache
+    { size: 100 },
+);
 
-            return hasAnyTokens;
-        default:
-            // Exhaustive check, all coin types in NETWORKS_WITH_TOKENS should be handled above
-            coin satisfies never;
+export const selectAccountTokenSymbol = (
+    state: AccountsRootState,
+    accountKey?: AccountKey,
+    tokenAddress?: TokenAddress,
+): TokenSymbol | null => {
+    const tokenInfo = selectAccountTokenInfo(state, accountKey, tokenAddress);
 
-            return false;
+    if (!tokenInfo) {
+        return null;
     }
+
+    // FIXME: This is the only place in the codebase where we change case of token symbol.
+    // The `toUpperCase()` operation is necessary because we are receiving wrongly formatted token symbol from connect.
+    // Can be removed at the moment when desktop issue https://github.com/trezor/trezor-suite/issues/8037 is resolved.
+    return tokenInfo.symbol.toUpperCase() as TokenSymbol;
 };
 
-export const selectNumberOfAccountTokensWithFiatRates = (
-    state: TokensRootState,
-    accountKey: string,
-) => {
+export const selectAccountTokenTransactions = memoizeWithArgs(
+    (
+        state: TransactionsRootState,
+        accountKey: AccountKey,
+        tokenAddress: TokenAddress,
+    ): WalletAccountTransaction[] =>
+        pipe(
+            selectAccountTransactions(state, accountKey),
+            A.map(transaction => ({
+                ...transaction,
+                tokens: transaction.tokens.map((tokenTransfer: TokenTransfer) => ({
+                    ...tokenTransfer,
+                    symbol: tokenTransfer.symbol,
+                })),
+            })),
+            A.filter(transaction =>
+                A.some(
+                    transaction.tokens,
+                    tokenTransfer => tokenTransfer.contract === tokenAddress,
+                ),
+            ),
+        ) as WalletAccountTransaction[],
+    { size: 500 },
+);
+
+const selectAllAccountTokens = (
+    state: AccountsRootState,
+    accountKey: AccountKey,
+): TokenInfoBranded[] => {
+    const account = selectAccountByKey(state, accountKey);
+    if (!account || !account.tokens) {
+        return [];
+    }
+
+    return account.tokens as TokenInfoBranded[];
+};
+
+export const selectAnyOfTokensIsKnown = (
+    state: TokenDefinitionsRootState & AccountsRootState,
+    accountKey: AccountKey,
+): boolean => {
+    // It may be temping to reuse selectAccountsKnownTokens.length but this is faster
+    const tokens = selectAllAccountTokens(state, accountKey);
     const account = selectAccountByKey(state, accountKey);
 
-    if (!account || !isCoinWithTokens(account.symbol)) return 0;
-
-    switch (account.symbol) {
-        case 'eth':
-            const tokensNumber = selectNumberOfEthereumAccountTokensWithFiatRates(
-                state,
-                accountKey,
-            );
-
-            return tokensNumber;
-        default:
-            // Exhaustive check, all coin types in NETWORKS_WITH_TOKENS should be handled above
-            account.symbol satisfies never;
-
-            return 0;
+    if (!account?.symbol) {
+        return false;
     }
+    const result = A.any(tokens, token => {
+        const isKnown = selectIsSpecificCoinDefinitionKnown(state, account.symbol, token.contract);
+
+        return isKnown;
+    });
+
+    return result;
+};
+
+const isNotZeroAmountTransfer = (tokenTransfer: TokenTransfer) =>
+    tokenTransfer.amount !== '' && tokenTransfer.amount !== '0';
+
+/** Is not a transaction with zero amount and no internal transfers. */
+const isNotEmptyTransaction = (transaction: WalletAccountTransaction) =>
+    transaction.amount !== '0' ||
+    (G.isArray(transaction.internalTransfers) && A.isNotEmpty(transaction.internalTransfers));
+
+const isTransactionWithTokenTransfers = (transaction: WalletAccountTransaction) =>
+    G.isArray(transaction.tokens) && A.isNotEmpty(transaction.tokens);
+
+const selectAccountTransactionsWithTokensWithFiatRates = memoizeWithArgs(
+    (
+        state: TransactionsRootState & TokenDefinitionsRootState,
+        accountKey: AccountKey,
+        areTokenOnlyTransactionsIncluded: boolean,
+    ): WalletAccountTransaction[] =>
+        pipe(
+            selectAccountTransactions(state, accountKey),
+            A.map(transaction => ({
+                ...transaction,
+                tokens: pipe(
+                    transaction?.tokens ?? [],
+                    A.filter(isNotZeroAmountTransfer),
+                    A.filter(token =>
+                        selectIsSpecificCoinDefinitionKnown(
+                            state,
+                            transaction.symbol,
+                            token.contract as TokenAddress,
+                        ),
+                    ),
+                    A.map((tokenTransfer: TokenTransfer) => ({
+                        ...tokenTransfer,
+                        symbol: tokenTransfer.symbol,
+                    })),
+                ) as TypedTokenTransfer[],
+            })),
+            A.filter(
+                transaction =>
+                    isNotEmptyTransaction(transaction) ||
+                    (areTokenOnlyTransactionsIncluded &&
+                        isTransactionWithTokenTransfers(transaction)),
+            ),
+        ) as WalletAccountTransaction[],
+    { size: 500 },
+);
+
+export const selectAccountOrTokenAccountTransactions = (
+    state: TransactionsRootState & TokenDefinitionsRootState,
+    accountKey: AccountKey,
+    tokenAddress: TokenAddress | null,
+    areTokenOnlyTransactionsIncluded: boolean,
+): WalletAccountTransaction[] => {
+    if (tokenAddress) {
+        return selectAccountTokenTransactions(state, accountKey, tokenAddress);
+    }
+
+    return selectAccountTransactionsWithTokensWithFiatRates(
+        state,
+        accountKey,
+        areTokenOnlyTransactionsIncluded,
+    ) as WalletAccountTransaction[];
+};
+
+export const selectAccountsKnownTokens = memoizeWithArgs(
+    (
+        state: AccountsRootState & TokenDefinitionsRootState,
+        accountKey: AccountKey,
+    ): TokenInfoBranded[] => {
+        const account = selectAccountByKey(state, accountKey);
+        if (!account || !isCoinWithTokens(account.symbol)) {
+            return [];
+        }
+
+        return selectFilterKnownTokens(
+            state,
+            account.symbol,
+            account.tokens ?? [],
+        ) as TokenInfoBranded[];
+    },
+    { size: 50 },
+);
+
+export const selectNumberOfAccountTokensWithFiatRates = (
+    state: TokenDefinitionsRootState & AccountsRootState,
+    accountKey: AccountKey,
+): number => {
+    const account = selectAccountByKey(state, accountKey);
+
+    if (!account || !isCoinWithTokens(account.symbol)) {
+        return 0;
+    }
+
+    const tokens = selectAccountsKnownTokens(state, accountKey);
+
+    return tokens.length;
+};
+
+export const selectHasDeviceAnyTokensForNetwork = (state: TokensRootState, coin: NetworkSymbol) => {
+    if (!isCoinWithTokens(coin)) {
+        return false;
+    }
+
+    const accounts = selectVisibleDeviceAccountsByNetworkSymbol(state, coin);
+
+    return A.any(accounts, account => {
+        const result = selectAnyOfTokensIsKnown(state, account.key);
+
+        return result;
+    });
 };
 
 export const selectAccountHasAnyKnownToken = (state: TokensRootState, accountKey: string) => {
     const account = selectAccountByKey(state, accountKey);
 
-    if (!account || !isCoinWithTokens(account.symbol)) return false;
-
-    switch (account.symbol) {
-        case 'eth':
-            const anyOfTokensIsKnown = selectAnyOfTokensIsKnown(state, accountKey);
-
-            return anyOfTokensIsKnown;
-        default:
-            // Exhaustive check, all coin types in NETWORKS_WITH_TOKENS should be handled above
-            account.symbol satisfies never;
-
-            return false;
+    if (!account || !isCoinWithTokens(account.symbol)) {
+        return false;
     }
+
+    const anyOfTokensIsKnown = selectAnyOfTokensIsKnown(state, accountKey);
+
+    return anyOfTokensIsKnown;
 };
