+++ suite-native/tokens/src/types.ts
@@ -1,15 +1,15 @@
-import { TokenTransfer } from '@trezor/blockchain-link';
+import { TokenTransfer as BlockchainLinkTokenTransfer } from '@trezor/blockchain-link';
 import {
     TokenAddress,
     TokenSymbol,
     WalletAccountTransaction as CommonWalletAccountTransaction,
 } from '@suite-common/wallet-types';
 
-export type EthereumTokenTransfer = Omit<TokenTransfer, 'symbol' | 'contract'> & {
+export type TypedTokenTransfer = Omit<BlockchainLinkTokenTransfer, 'symbol' | 'contract'> & {
     symbol: TokenSymbol;
     contract: TokenAddress;
 };
 
 export type WalletAccountTransaction = Omit<CommonWalletAccountTransaction, 'tokens'> & {
-    tokens: EthereumTokenTransfer[];
+    tokens: TypedTokenTransfer[];
 };
