+++ suite-native/tokens/src/utils.ts
@@ -2,13 +2,13 @@ import { G, S } from '@mobily/ts-belt';
 
 import { NetworkSymbol } from '@suite-common/wallet-config';
 
-export const getEthereumTokenName = (tokenName?: string) => {
+export const getTokenName = (tokenName?: string) => {
     if (G.isNullable(tokenName) || S.isEmpty(tokenName)) return 'Unknown token';
 
     return tokenName;
 };
 
-export const NETWORK_SYMBOLS_WITH_TOKENS = ['eth'] satisfies Array<NetworkSymbol>;
+export const NETWORK_SYMBOLS_WITH_TOKENS = ['eth', 'pol', 'bnb'] satisfies Array<NetworkSymbol>;
 export type NetworkSymbolWithTokens = (typeof NETWORK_SYMBOLS_WITH_TOKENS)[number];
 
 export const isCoinWithTokens = (symbol: NetworkSymbol): symbol is NetworkSymbolWithTokens => {
