+++ suite-native/tokens/tsconfig.json
@@ -15,9 +15,6 @@
             "path": "../../suite-common/wallet-types"
         },
         {
-            "path": "../../suite-common/wallet-utils"
-        },
-        {
             "path": "../../packages/blockchain-link"
         }
     ]
