+++ suite-native/transactions/src/components/TransactionsList/TransactionList.tsx
@@ -15,7 +15,7 @@ import { AccountKey, TokenAddress } from '@suite-common/wallet-types';
 import { groupTransactionsByDate, isPending, MonthKey } from '@suite-common/wallet-utils';
 import { Box, Loader } from '@suite-native/atoms';
 import {
-    EthereumTokenTransfer,
+    TypedTokenTransfer,
     selectAccountOrTokenAccountTransactions,
     WalletAccountTransaction,
 } from '@suite-native/tokens';
@@ -49,7 +49,7 @@ type RenderTransactionItemParams = {
     isLast: boolean;
 };
 
-type EthereumTokenTransferWithTx = EthereumTokenTransfer & {
+type TypedTokenTransferWithTx = TypedTokenTransfer & {
     originalTransaction: WalletAccountTransaction;
 };
 
@@ -57,12 +57,12 @@ type RenderTokenTransferItemParams = Omit<
     RenderTransactionItemParams,
     'areTokensIncluded' | 'item'
 > & {
-    item: EthereumTokenTransferWithTx;
+    item: TypedTokenTransferWithTx;
     txid: string;
 };
 
 type TransactionListItem =
-    | (EthereumTokenTransferWithTx | MonthKey)
+    | (TypedTokenTransferWithTx | MonthKey)
     | (WalletAccountTransaction | MonthKey);
 
 const sectionListStyle = prepareNativeStyle(utils => ({
@@ -232,7 +232,7 @@ export const TransactionList = ({
                                 ({
                                     ...tokenTransfer,
                                     originalTransaction: transaction,
-                                }) as EthereumTokenTransferWithTx,
+                                }) as TypedTokenTransferWithTx,
                         ),
                 ),
             ]);
@@ -247,7 +247,7 @@ export const TransactionList = ({
     const renderItem = useCallback(
         ({ item, index }: { item: TransactionListItem; index: number }) => {
             if (typeof item === 'string') {
-                return renderSectionHeader({ section: { monthKey: item } });
+                return renderSectionHeader({ section: { monthKey: item as MonthKey } });
             }
             const isFirstInSection = typeof data.at(index - 1) === 'string';
             const isLastInSection =
@@ -255,7 +255,7 @@ export const TransactionList = ({
 
             const getIsTokenTransfer = (
                 itemForCheck: TransactionListItem,
-            ): itemForCheck is EthereumTokenTransferWithTx => 'originalTransaction' in itemForCheck;
+            ): itemForCheck is TypedTokenTransferWithTx => 'originalTransaction' in itemForCheck;
 
             return getIsTokenTransfer(item)
                 ? renderTokenTransferItem({
