+++ suite-native/transactions/src/components/TransactionsList/TransactionListItemContainer.tsx
@@ -21,7 +21,7 @@ import {
     TransactionsRootState,
 } from '@suite-common/wallet-core';
 import { NetworkSymbol } from '@suite-common/wallet-config';
-import { EthereumTokenTransfer } from '@suite-native/tokens';
+import { TypedTokenTransfer } from '@suite-native/tokens';
 import { Color } from '@trezor/theme';
 import { Translation } from '@suite-native/intl';
 import { TokenDefinitionsRootState } from '@suite-common/token-definitions';
@@ -37,7 +37,7 @@ type TransactionListItemContainerProps = RequireExactlyOne<
         isFirst?: boolean;
         isLast?: boolean;
         networkSymbol: NetworkSymbol;
-        tokenTransfer: EthereumTokenTransfer;
+        tokenTransfer: TypedTokenTransfer;
         transactionType: TransactionType;
     },
     'networkSymbol' | 'tokenTransfer'
